Ver 3.0.1.7
Added placeholders for RNG heroes, added functionality for Taint, Putrid, Defile, Neil. Made some variables double to increase precision for high damage WB and made WB health 64 bit. Note that absorb won't currently work properly if a lane has two absorb heroes in it.

+++ battleLogic.cpp
@@ -8,7 +8,7 @@ bool isBetter(Monster * a, Monster * b, bool considerAbilities) {
     if (a->element == b->element) {
         return (a->damage >= b->damage) && (a->hp >= b->hp);
     } else { // a needs to be better than b even when b has elemental advantage, or a is at disadvantage
-        return !considerAbilities && (a->damage >= (int16_t) ((float) b->damage * elementalBoost)) && (a->hp >= (int16_t) ((float) b->hp * elementalBoost));
+        return !considerAbilities && (a->damage >= (int16_t) ((double) b->damage * elementalBoost)) && (a->hp >= (int16_t) ((double) b->hp * elementalBoost));
     }
 }
